--- rosbag2_transport/include/rosbag2_transport/play_options_from_node_params.hpp
@@ -1,29 +1,0 @@
-// Copyright 2023 Open Source Robotics Foundation, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef ROSBAG2_TRANSPORT__PLAY_OPTIONS_FROM_NODE_PARAMS_HPP_
-#define ROSBAG2_TRANSPORT__PLAY_OPTIONS_FROM_NODE_PARAMS_HPP_
-
-#include <memory>
-
-#include "rclcpp/node.hpp"
-#include "rosbag2_transport/play_options.hpp"
-
-namespace rosbag2_transport
-{
-rosbag2_transport::PlayOptions
-get_play_options_from_node_params(std::shared_ptr<rclcpp::Node> node);
-}  // namespace rosbag2_transport
-
-#endif  // ROSBAG2_TRANSPORT__PLAY_OPTIONS_FROM_NODE_PARAMS_HPP_
